Move test setup for AST tests into AstTestSetup
